IN PROGRESS - issue NEXUS-2501: User groups/roles are not migrated. https://issues.sonatype.org/browse/NEXUS-2501
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus-plugins/branches/migration-plugin-1.1.x@359 c41184c3-3edd-4803-b646-023156917d6f

+++ nexus-migration-plugin-test-harness/src/test/java/org/sonatype/nexus/plugins/migration/nexus1523/Nexus1523ImportToAnExistingRepoTest.java
@@ -2,10 +2,12 @@ package org.sonatype.nexus.plugins.migration.nexus1523;
 
 import java.io.File;
 import java.io.FileReader;
+import java.io.IOException;
 
 import org.apache.commons.io.FileUtils;
 import org.codehaus.plexus.util.IOUtil;
 import org.junit.Assert;
+import org.junit.BeforeClass;
 import org.junit.Test;
 import org.sonatype.nexus.plugin.migration.artifactory.dto.MigrationSummaryDTO;
 import org.sonatype.nexus.plugins.migration.AbstractMigrationIntegrationTest;
@@ -15,18 +17,27 @@ public class Nexus1523ImportToAnExistingRepoTest
     extends AbstractMigrationIntegrationTest
 {
 
+    @BeforeClass
+    public static void deleteLog()
+        throws IOException
+    {
+        File logFile = new File( "./target/logs/migration.log" );
+        if ( logFile.exists() )
+        {
+            FileUtils.forceDelete( logFile );
+        }
+    }
+
     @Test
     public void importToAnExistingRepo()
         throws Exception
     {
-        File logFile = new File( "./target/logs/migration.log" );
-        FileUtils.forceDelete( logFile );
-
         MigrationSummaryDTO migrationSummary = prepareMigration( getTestFile( "artifactoryBackup.zip" ) );
         commitMigration( migrationSummary );
         TaskScheduleUtil.waitForTasks( 40 );
         Thread.sleep( 2000 );
 
+        File logFile = new File( "./target/logs/migration.log" );
         Assert.assertTrue( "Migration log file not found", logFile.isFile() );
 
         String log = IOUtil.toString( new FileReader( logFile ) );
